Update ShortCutAgents.py
The import of env was incorrect as the Environment module has no env class or method, made it an argument for the agents initializations and opened doors to reducing arguments size by using the env methods. (View comments in init magic method)

Important to remember, in nStepSARSAAgent train method we send cumulative reward (int) to update as reward (np.array), this will lead to issues during testing and experimenting.